# performance: Maintain and use a dependency graph