--- a/.lsp/config.edn
+++ b/.lsp/config.edn
@@ -1,3 +1,4 @@
 {:linters {:clj-kondo {:ns-exclude-regex "sample-test.*"}}
  :source-aliases #{:dev :test :debug}
- :use-metadata-for-privacy? true}
+ :use-metadata-for-privacy? true
+ #_#_:experimental {:dep-graph-queries true}}