--- a/Makefile
+++ b/Makefile
@@ -49,6 +49,7 @@ native-cli:
 
 test: classes
 	cd lib && clojure -M:test
+	cd lib && CLOJURE_LSP_USE_DEP_GRAPH="true" clojure -M:test
 	cd cli && clojure -M:test
 
 pod-test: classes