--- a/Makefile
+++ b/Makefile
@@ -51,6 +51,7 @@ test: classes
 	cd lib && clojure -M:test
 	cd lib && CLOJURE_LSP_USE_DEP_GRAPH="true" clojure -M:test
 	cd cli && clojure -M:test
+	cd cli && CLOJURE_LSP_USE_DEP_GRAPH="true" clojure -M:test
 
 pod-test: classes
 	cd cli && clojure -M:pod-test