--- a/cli/integration-test/sample-test/.lsp/config.edn
+++ b/cli/integration-test/sample-test/.lsp/config.edn
@@ -1,3 +1,4 @@
 {:linters {:clj-kondo {:ns-exclude-regex ""}}
  :log-path "clojure-lsp.integration-test.out"
- :api {:exit-on-errors? false}}
+ :api {:exit-on-errors? false}
+ #_#_:experimental {:dep-graph-queries true}}