--- a/lib/src/clojure_lsp/dep_graph.clj
+++ b/lib/src/clojure_lsp/dep_graph.clj
@@ -87,7 +87,6 @@
 
 ;; :langs is a set of the langs used by this file.
 
-
 ;;;; DB Maintenance
 
 (def ^:private empty-multiset {})