… CHANGELOG.md
… docs/settings.md
… lib/src/clojure_lsp/feature/clean_ns.clj
… lib/src/clojure_lsp/feature/code_lens.clj
… lib/src/clojure_lsp/feature/completion.clj
… lib/src/clojure_lsp/feature/diagnostics.clj
… lib/src/clojure_lsp/feature/drag.clj
… lib/src/clojure_lsp/feature/java_interop.clj
… lib/src/clojure_lsp/feature/move_form.clj
… lib/src/clojure_lsp/feature/test_tree.clj
… lib/src/clojure_lsp/feature/workspace_symbols.clj
… lib/src/clojure_lsp/handlers.clj
… lib/src/clojure_lsp/internal_api.clj
… lib/src/clojure_lsp/kondo.clj
… lib/src/clojure_lsp/queries.clj
… lib/src/clojure_lsp/refactor/transform.clj
… lib/test/clojure_lsp/dep_graph_test.clj
… lib/test/clojure_lsp/features/file_management_test.clj
… lib/test/clojure_lsp/queries_test.clj
… lib/test/clojure_lsp/refactor/transform_test.clj
… lib/test/clojure_lsp/settings_test.clj