Maintains a dependency graph and file meta-data to be able to more efficiently filter the clj-kondo analysis.

The dep-graph is a map from namespaces to the namespaces they depend on (their dependencies) and the namespaces which depend on them (their dependents). Along with file meta-data, which records which namespaces a file defines, this makes it easy to quickly filter to a much narrower part of the analysis. For example, when finding references of a var, you need to look only in the analysis of the var's namespace's dependent's files.

Since so many of the queries inspected every element of every file in the whole analysis, performance was tied to the size of the project. Now, in many cases performance is tied instead to the combined sizes of a file and its dependencies or dependents.

This makes many parts of the code faster, but is most noticeable in a few places.

In metabase, a very large project, find-references typically took 100-120ms, which was the time to look at every var usage. The time hasn't improved when finding references to widely used vars like `defn` because most namespaces depend on clojure.core, so we still have to search through all the var usages. But for namespaces that have fewer dependents, find-references is much faster, often around 20ms and sometimes less than 10ms.

Similarly, find-definition went from 60ms to 20ms.

But find-references and find-definition are both invoked by the user relatively rarely. The big payoff is in completion, which is invoked regularly, sometimes on every keystroke.

In metabase, before:

```
completing 'comp' "Elapsed time: 330.945051 msecs"
completing 'config/e' "Elapsed time: 251.061651 msecs"
completing '::lmf/d' "Elapsed time: 282.454571 msecs"
```

After:

```
completing 'comp' "Elapsed time: 23.971551 msecs"
completing 'config/e' "Elapsed time: 13.562191 msecs"
completing '::lmf/d' "Elapsed time: 96.987585 msecs"
```

These are just using `time`, not something more rigorous, but the effect is still apparent. In a medium size project, clojure-lsp, before:

```
completing 'comp' "Elapsed time: 105.267534 msecs"
completing 'shared/f' "Elapsed time: 94.15348 msecs"
completing '::coercer/c' "Elapsed time: 103.5377 msecs"
```

After:

```
completing 'comp' "Elapsed time: 8.83796 msecs"
completing 'shared/f' "Elapsed time: 9.592488 msecs"
completing '::coercer/c' "Elapsed time: 32.093406 msecs"
```

The biggest improvement in completion comes from minimizing time in `with-elements-from-alias` and `cursor-full-ns?`, finding aliases and namespace names, respectively. Previously these were a major contributor to completion time. Reducing time in these speeds up even things which can't use the dep graph, like keyword completion.

Notes on automated tests
========

You can run the test suites with dep-graph-queries turned on by un-commenting the `:dep-graph-queries` setting in `common-test/src/clojure_lsp/test_helper.clj` and `cli/integration-test/sample-test/.lsp/config.edn` for unit and integration tests respectively.

A few lower-level unit tests about analysis-narrowing return different results depending on whether `:dep-graph-queries` is on. This is expected, and proves that the feature is working. But otherwise this optimization should be transparent to the rest of the code, so other tests haven't changed.

It would be sort of nice to have a "test matrix", so that the full suite was run twice, once with `:dep-graph-queries` on and once with it off. But assuming that's a pain to set up, I'm OK with running them both manually until this experiment ends.

Notes for beta testers
========

With a build from this branch, you can test this feature by adding `:experimental {:dep-graph-queries true}` to your `.lsp/config.edn`. (If you work on clojure-lsp itself, that setting exists but is commented out.)

You can toggle the feature on and off by changing the setting, without restarting the server. (You have to wait at least 1 second for the new settings to be picked up. Sometimes this seems flaky.) This is useful to compare the results of certain actions and to compare performance.

Effects on all users
========

This patch will have some effect on all users, even if they aren't using the experimental feature.

In order for the dep-graph to keep track of the relationships between files in external dependencies (i.e. those outside of the project) it needs the clj-kondo analysis to include `:namespace-usages`. There isn't a way to turn off that part of clj-kondo so we've always received that when running the external analysis, but we used to dissoc it before putting it in the db and before caching it on disk. Now we need to keep that analysis around, including in the disk cache. For this reason, the db/version has been bumped, meaning that the external dependencies of all users' projects will be re-scanned, and that the size of cache files will be somewhat larger. (I haven't measured the difference in file size.)

Additionally, the dep-graph is built and maintained for all users, regardless of whether they're using its data. In metabase this adds about 1.5s to startup. In clojure-lsp it adds about 500ms. Ongoing maintenance of the dep-graph after startup is negligible. This approach means anyone can toggle the feature on without restarting the server (see above "Notes for beta testers"). But if the additional startup time isn't acceptable, this approach can be re-thought.

Notes for PR reviewers
===
* When looking at the diffs, I suggest hiding whitespace changes. 
* I'd love to have testers for this branch before its merged. After its merged, assuming it is, I hope we can appeal to the community for additional testing.
* And if you really want to stress test the feature, disable your editors' completion caching 😈 .

- [x] This performance improvement was discussed in #990
- [x] I added a new entry to [CHANGELOG.md](https://github.com/clojure-lsp/clojure-lsp/blob/master/CHANGELOG.md)
- [x] I updated documentation if applicable (`docs` folder)