@mainej 2022-06-16
Thanks for the feedback @ericdallo and @snoe! I'm busy with some other stuff, but later I'll take some time to process it all.

I wanted to give a quick update though. Originally I reported that find-definition went from 60ms to 20ms, but that seemed surprisingly long. I also noticed that find-definition was dominating the time in certain completion scenarios. I tracked down the problem and now find-definition is essentially instantaneous, 1ms in many cases.

Also, in the initial PR report I neglected a particular completion which is still slow. In metabase, if you type `met` you get almost 600 completions. Almost all of them are suggestions to complete to a namespace (there are a lot of namespaces that start with `met`). That takes 350-400ms. Looking at a flamegraph, 95% of that time is in `completion/with-elements-from-alias`, either in `feature.add-missing-libspec/add-known-alias` or in `refactor.transform/result`. We're calculating not just the completion, but also the edit that would add the namespace to the `:require`s. I think we could postpone that edit calculation to `completion/resolve-item`, so that we calculate only one edit, and only after the user chooses it. I think that'd be a nice optimization, but would rather postpone it to a separate PR.