@mainej 2022-06-18
@snoe, I agree, performance tests—especially performance tests that integrate with real projects—would be handy. They'd be useful for developing this kind of feature but also for noticing regressions caused by non-performance related features. We've been collecting ideas in #909 for awhile. Would you mind adding your thoughts there about checking out projects at known versions? I really like that idea. I keep meaning to work on #909, but then get distracted by actually fixing performance problems :)

@ericdallo, I'll look at creating test matrices with an env var. For some reason the GitHub actions daunt me, so if this would be easy for you, I'd welcome help.

Also, you've figured out exactly what would be needed to limit the impact on all users: 
1. Figure out how to avoid the db version bump and/or conditionally dissoc the :namespace-usages.
2. Only conditionally run `dep-graph/refresh-analysis`.

Do you want me to look into that more, or leave it like it is?