@ericdallo 2022-06-18
I think we can leave it like this for now, we are aware of what will be affected and we intend to take that risk in change of consistency.

I can help with the test matrix, do a PR on this one or even add commits directly to this branch.

I was thinking in release a new clojure-lsp version on Monday, do you think makes sense leave this PR to not this one but next release? I still need to review code and test it manually too