@mainej 2022-06-19
> I can help with the test matrix, do a PR on this one or even add commits directly to this branch.

I'd appreciate that. Commits on this branch would be fine. I'll avoid force-pushing.

> I was thinking in release a new clojure-lsp version on Monday, do you think makes sense leave this PR to not this one but next release? I still need to review code and test it manually too

Yeah, probably best to postpone it to the next release cycle. No need to rush it through.