@mainej 2022-06-20
@ericdallo that's simpler than I was envisioning. I thought you'd use GitHub Action [env vars](https://docs.github.com/en/actions/learn-github-actions/environment-variables) and [test matrices](https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs). But this way is nice and simple, and will get around messing with the `config.edn` files or the initial value of `:settings`. I kind of think we should run the CLI and integration tests with `CLOJURE_LSP_USE_DEP_GRAPH` too. WDYT?

Also... this is confusing... [This test](https://github.com/clojure-lsp/clojure-lsp/runs/6957662332?check_suite_focus=true) had an ERROR when running with the dep-graph, but the suite passed. Any idea why?