@mainej 2022-06-20
> queries should affect lib only

Thanks for adding the CLI test... Queries are used in the CLI too, by `rename` and `clean-ns`.

> your link points just to a succeeded build

True, the build succeeded, but it shouldn't have. Search the test output for "ERROR".