@ericdallo 2022-06-20
because it's a exception in a async go loop, the test doesn't catch this kind of exception