@mainej 2022-06-20
So kaocha notices that an exception was thrown but doesn't fail the test?