@ericdallo 2022-06-20
> So kaocha notices that an exception was thrown but doesn't fail the test?

Yeah, I think so, because it's a print to stdout, not really a `(is)` fail or a exception on main test thread