@ericdallo 2022-07-01
FYI: tested on one of Nubank biggest clojure only projects (> 180K loc) and it's awesome to see something like:
```
:maintain-dep-graph 262ms
```