@mainej 2022-07-01
> `:maintain-dep-graph 2119ms`

I'm pleased it doesn't slow down startup _too_ much. I think it's fast enough for now, but it's still a little longer than I'd like. A follow up PR will make it a tiny bit faster. But the real bummer is that whenever you run internal analysis, the code has to update the dep-graph twice, ending up with the same results both times. So, if you look at the logs again, I'd expect you to  see the 2 seconds of maintenance appear twice.

We've had this double calculation problem for a long time—normalization happens twice too—but now we have it with dep-graph maintenance as well. The reason is custom-lint-fn, which is why I was asking about it a few days ago. The custom lint fns receive raw clj-kondo analysis. To find unused public vars they need to query the db with that additional analysis, so they have to normalize it, merge it, and update the dep-graph. They do this speculatively, without actually touching the db atom. With this temp db they figure out the custom diagnostics, then return them to clj-kondo via `reg-finding!`. `reg-finding!` filters a few of them out based on the "ignores" that it knows about, and then returns the diagnostics to us along with the other diagnostics it found. It also returns the raw analysis again. Then we have to take that raw analysis, re-normalize, re-update the dep-graph, and finally swap everything into the db. In pictures, it looks like this:

```
clojure-lsp: change -> 
kondo: pass raw analysis to custom-lint-fn -> 
clojure-lsp: normalization+dep-graph -> (speculatively)
clojure-lsp: custom lint -> 
kondo: filter custom diagnostics by level and ignores ->
kondo: return raw analysis+full diagnostics ->
clojure-lsp: normalization+dep-graph -> (duplicate work!)
clojure-lsp: swap db
```

What I'd like to do is remove the custom lint fns and instead do this:

```
clojure-lsp: change -> 
kondo: return raw analysis+kondo diagnostics -> 
clojure-lsp: normalization+dep-graph -> 
clojure-lsp: custom lint -> 
kondo: filter custom diagnostics by level and ignores ->
clojure-lsp: concat kondo and custom diagnostics ->
clojure-lsp: swap db
```
The only thing making this difficult is the "ignores" filtering that `reg-finding!` does. That filtering isn't public, and it depends on the full `:ctx`, which holds the "ignores" data. So, in theory we could do it with a little refactoring to expose the filtering publicly, but we'd also need kondo to return the `:ctx`, so that we could use it _after_ `clj-kondo/run!` returns. We could work with @borkdude on the details, but if I were him I wouldn't be excited about leaking all the implementation detail of `:ctx`.

Again, it's not a huge problem—the duplicate work only amounts to a few seconds at startup and a few milliseconds as individual files are updated—but it still seems wasteful.